Nest rules in /preprocessor. Move dataValidator/rules/topstories.js to /lib.
